+++ app/src/main/java/eu/kanade/tachiyomi/data/track/kitsu/KitsuUtils.kt
@@ -3,20 +3,38 @@ package eu.kanade.tachiyomi.data.track.kitsu
 import eu.kanade.tachiyomi.data.database.models.anime.AnimeTrack
 import eu.kanade.tachiyomi.data.database.models.manga.MangaTrack
 
-fun MangaTrack.toApiStatus() = when (status) {
-    Kitsu.READING -> "current"
-    Kitsu.COMPLETED -> "completed"
-    Kitsu.ON_HOLD -> "on_hold"
-    Kitsu.DROPPED -> "dropped"
-    Kitsu.PLAN_TO_READ -> "planned"
-    else -> throw Exception("Unknown status")
+fun MangaTrack.toKitsuApiStatus() = when (status) {
+    Kitsu.READING -> "CURRENT"
+    Kitsu.COMPLETED -> "COMPLETED"
+    Kitsu.ON_HOLD -> "ON_HOLD"
+    Kitsu.DROPPED -> "DROPPED"
+    Kitsu.PLAN_TO_READ -> "PLANNED"
+    else -> throw Exception("Unknown status: $status")
 }
 
-fun AnimeTrack.toApiStatus() = when (status) {
-    Kitsu.WATCHING -> "current"
-    Kitsu.COMPLETED -> "completed"
-    Kitsu.ON_HOLD -> "on_hold"
-    Kitsu.DROPPED -> "dropped"
-    Kitsu.PLAN_TO_WATCH -> "planned"
-    else -> throw Exception("Unknown status")
+fun AnimeTrack.toKitsuApiStatus() = when (status) {
+    Kitsu.WATCHING -> "CURRENT"
+    Kitsu.COMPLETED -> "COMPLETED"
+    Kitsu.ON_HOLD -> "ON_HOLD"
+    Kitsu.DROPPED -> "DROPPED"
+    Kitsu.PLAN_TO_WATCH -> "PLANNED"
+    else -> throw Exception("Unknown status: $status")
+}
+
+fun String.toKitsuLocalStatus() = when (this) {
+    "CURRENT" -> Kitsu.READING
+    "COMPLETED" -> Kitsu.COMPLETED
+    "ON_HOLD" -> Kitsu.ON_HOLD
+    "DROPPED" -> Kitsu.DROPPED
+    "PLANNED" -> Kitsu.PLAN_TO_READ
+    else -> throw Exception("Unknown status: $this")
+}
+
+fun String.toKitsuLocalStatusAnime() = when (this) {
+    "CURRENT" -> Kitsu.WATCHING
+    "COMPLETED" -> Kitsu.COMPLETED
+    "ON_HOLD" -> Kitsu.ON_HOLD
+    "DROPPED" -> Kitsu.DROPPED
+    "PLANNED" -> Kitsu.PLAN_TO_WATCH
+    else -> throw Exception("Unknown status: $this")
 }